+++ include/mruby_integration/textures.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_textures(mrb_state*);